+++ src/@cli/@views/OutputViews/LottoListOutputView.js
@@ -1,4 +1,4 @@
-import Console from "../util/Console.js";
+import Console from "../../util/Console.js";
 
 class LottoListOutputView {
     static render (lottoList) {
