+++ src/domain/LottoList/LottoList.js
@@ -19,7 +19,7 @@ export class LottoList {
         return this.#lottoList.filter(lotto => lotto.winningRank === winningRank).length;
     };
 
-    getLottoResultSummary({order= 'ASC'}){
+    getLottoResultSummary({order}={order: 'ASC'}){
         const resultSummary = LOTTO_INFO.WINNING_CONDITION.map(CONDITION => ({
             match: CONDITION.MATCH,
             bonus: CONDITION.BONUS,
